fix: Unnecessary Socket disconnect/reconnect in ParticipantSession fixed
