+++ frontend/src/context/ParticipantSessionContext.jsx
@@ -5,7 +5,9 @@ import React, {
   useMemo,
   useState,
   useRef,
+  useCallback,
 } from "react";
+import { io } from "socket.io-client";
 
 export const STORAGE_KEY = "participantSession";
 
@@ -23,6 +25,7 @@ export const ParticipantSessionProvider = ({ children }) => {
   });
   // const [activePoll, setActivePoll] = useState(null);
   const socketRef = useRef(null);
+  const [socketReady, setSocketReady] = useState(false);
   const [broadcastMsg, setBroadcastMsg] = useState(null);
   const [participantCount, setParticipantCount] = useState(1);
   const [selectedOption, setSelectedOption] = useState(null);
@@ -48,14 +51,67 @@ export const ParticipantSessionProvider = ({ children }) => {
     }
   }, [sessionData]);
 
-  const clearSession = () => {
+  const clearSession = useCallback(() => {
+    // remove persisted session and active poll immediately to avoid
+    // consumers re-hydrating from storage and re-creating sockets.
+    try {
+      sessionStorage.removeItem(STORAGE_KEY);
+      sessionStorage.removeItem("activePoll");
+    } catch {}
+
     setSessionData(null);
     setActivePoll(null);
-  };
+  }, []);
+
+  // Socket lifecycle - create once per provider when sessionData exists
+  const SOCKET_URL = (import.meta.env?.VITE_BACKEND_BASE_URL || "http://localhost:3000") + "/sessions";
+  useEffect(() => {
+    // only create when we have session data and no socket
+    if (!sessionData || socketRef.current) return;
+
+    const token = localStorage.getItem("accessToken");
+    const socket = io(SOCKET_URL, {
+      withCredentials: true,
+      extraHeaders: token ? { Authorization: `Bearer ${token}` } : undefined,
+    });
+    socketRef.current = socket;
+    setSocketReady(true);
+
+    // emit join on connect
+    socket.on("connect", () => {
+      try {
+        socket.emit("join-session", {
+          code: sessionData.joinCode,
+          participantId: sessionData.participantId,
+          role: "student",
+        });
+      } catch {}
+    });
+
+    // on unmount or when sessionData is cleared, leave and disconnect
+    return () => {
+      try {
+        if (socketRef.current) {
+          socketRef.current.emit("leave-session", {
+            code: sessionData.joinCode,
+            participantId: sessionData.participantId,
+          });
+        }
+      } catch {}
+      try {
+        socket.disconnect();
+      } catch {}
+      socketRef.current = null;
+      setSocketReady(false);
+    };
+    // we intentionally only want to run when sessionData becomes available or cleared
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [sessionData]);
 
   const value = useMemo(
     () => ({
       socketRef,
+      socketReady,
       clearSession,
 
       sessionData,
@@ -82,7 +138,7 @@ export const ParticipantSessionProvider = ({ children }) => {
       pollId, 
       setPollId
     }),
-    [sessionData, activePoll]
+    [sessionData, activePoll, socketReady]
   );
 
   return (
